feat(mobileApp): Complete TypeScript migration for entry points and helpers
