+++ apps/mobileApp/package.json
@@ -5,9 +5,10 @@
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
-    "lint": "eslint .",
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "start": "react-native start",
-    "test": "jest"
+    "test": "jest",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
     "@appandflow/react-native-google-autocomplete": "^2.1.0",
@@ -19,10 +20,10 @@
     "@react-native-google-signin/google-signin": "^15.0.0",
     "@react-native/codegen": "^0.78.1",
     "@react-native/gradle-plugin": "^0.78.1",
-    "@react-navigation/bottom-tabs": "^7.3.3",
-    "@react-navigation/native": "^7.0.19",
-    "@react-navigation/native-stack": "^7.3.3",
-    "@react-navigation/stack": "^7.2.3",
+    "@react-navigation/bottom-tabs": "^7.4.7",
+    "@react-navigation/native": "^7.1.17",
+    "@react-navigation/native-stack": "^7.3.26",
+    "@react-navigation/stack": "^7.4.8",
     "@reduxjs/toolkit": "^2.6.1",
     "aws-sdk": "2.1692.0",
     "axios": "^1.8.4",
@@ -57,8 +58,8 @@
     "react-native-push-notification": "^8.1.1",
     "react-native-raw-bottom-sheet": "^3.0.0",
     "react-native-reanimated": "^3.19.0",
-    "react-native-safe-area-context": "^5.3.0",
-    "react-native-screens": "^4.10.0",
+    "react-native-safe-area-context": "3.4.1",
+    "react-native-screens": "^3.32.0",
     "react-native-shimmer-placeholder": "^2.0.9",
     "react-native-splash-screen": "^3.3.0",
     "react-native-status-bar-height": "^2.6.0",
@@ -84,12 +85,11 @@
     "@react-native/typescript-config": "0.78.1",
     "@types/jest": "^29.5.13",
     "@types/react": "^19.1.12",
-    "@types/react-native": "^0.73.0",
     "@types/react-test-renderer": "^19.0.0",
     "@typescript-eslint/eslint-plugin": "^8.43.0",
     "@typescript-eslint/parser": "^8.43.0",
     "babel-plugin-module-resolver": "^5.0.2",
-    "eslint": "^8.19.0",
+    "eslint": "^8.57.0",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
     "react-test-renderer": "19.0.0",
